fix startPosition
라운드마다 모든 유저(보스 포함)의 좌표값을 초기화하여 보스가 죽은 유저 옆에서 대기하다 바로 공격이 불가능하도록 수정
