+++ src/handlers/enumTyps.ts
@@ -54,6 +54,9 @@ export const CardType = Object.freeze({
   ROGUE_EXTENDED_SKILL: 107,
   WARRIOR_EXTENDED_SKILL: 108,
   MAGICIAN_FINAL_SKILL: 109,
+  ARCHER_FINAL_SKILL: 110,
+  ROGUE_FINAL_SKILL: 111,
+  WARRIOR_FINAL_SKILL: 112,
   BOSS_BASIC_SKILL: 113,
   BASIC_HP_POTION: 201,
   BASIC_MP_POTION: 202,
